Make ruff happy

+++ av/sidedata/encparams.pyx
@@ -1,5 +1,5 @@
 cimport libav as lib
-from libc.stdint cimport uint8_t, int32_t
+from libc.stdint cimport int32_t, uint8_t
 
 from enum import IntEnum
 
@@ -85,7 +85,7 @@ cdef class VideoEncParams(SideData):
         Only for MPEG2 and H264 encoded videos.
         """
         import numpy as np
-        
+
         cdef int mb_h = (self.frame.ptr.height + 15) // 16
         cdef int mb_w = (self.frame.ptr.width + 15) // 16
         cdef int nb_mb = mb_h * mb_w
